backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
pydantic==2.7.0
groq==0.9.0
tavily-python==0.3.3
pypdf2==3.0.1
python-docx==1.1.0
sentence-transformers==2.7.0
numpy==1.26.4
supabase==2.3.0
httpx==0.27.0
openpyxl==3.1.2
python-pptx==0.6.23
